feat: 홈 화면 UI 개선 및 컴포넌트 분리
- `HomeScreen`:
    - `ModernBottomBar` (새로운 하단 네비게이션 바) 추가 및 선택된 탭 상태 관리
    - `RealtimeGraph`를 사용하여 실시간 소음 데이터 시각화
    - 기존에 `HomeScreen` 내부에 있던 `StatWidget`, `AIAnalysisCard`, `QuickActionsGrid`, `PulsingDot`, `MiniWaveform`, `FloatingParticles`, `ControlButton` 컴포넌트들을 별도 파일로 분리
    - `noiseLevelHistory`를 통해 최근 50개의 소음 데이터 관리 및 `RealtimeGraph`에 전달

- `component/`:
    - `AIAnalysisCard.kt`: AI 기반 소음 분석 결과 표시 카드 컴포넌트
    - `ControlButton.kt`: 녹음 시작/중지/일시정지 등의 제어 버튼 컴포넌트
    - `FloatingParticles.kt`: 녹음 중 화면에 표시되는 플로팅 파티클 효과 컴포넌트 및 소음 레벨에 따른 색상 결정 함수(`getNoiseColor`)
    - `MiniWaveform.kt`: 상단 바에 표시되는 작은 실시간 파형 애니메이션 컴포넌트
    - `ModernBottomBar.kt`: iOS 스타일의 하단 네비게이션 바 컴포넌트 (`TabItem` 포함)
    - `PulsingDot.kt`: 녹음 상태를 나타내는 맥동하는 점 애니메이션 컴포넌트
    - `QuickActionsGrid.kt`: 빠른 실행 기능을 위한 그리드 레이아웃 컴포넌트 (`QuickActionTile` 포함)
    - `StatWidget.kt`: 평균/최저/최고 소음 통계 표시 위젯 컴포넌트
