+++ composeApp/src/commonMain/kotlin/net/lateinit/noiseguard/presentation/ui/component/ModernBottomBar.kt
@@ -1,0 +1,174 @@
+package net.lateinit.noiseguard.presentation.ui.component
+
+import androidx.compose.animation.animateColorAsState
+import androidx.compose.animation.core.Spring
+import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.animation.core.spring
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Assessment
+import androidx.compose.material.icons.filled.History
+import androidx.compose.material.icons.filled.Home
+import androidx.compose.material.icons.filled.Settings
+import androidx.compose.material.icons.outlined.Analytics
+import androidx.compose.material.icons.outlined.History
+import androidx.compose.material.icons.outlined.Settings
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.scale
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import net.lateinit.noiseguard.presentation.theme.Gray700
+import net.lateinit.noiseguard.presentation.theme.Primary700
+
+@Composable
+fun ModernBottomBar(
+    selectedTab: Int,
+    onTabSelected: (Int) -> Unit,
+    isRecording: Boolean
+) {
+    val tabs = listOf(
+        TabItem(Icons.Filled.Home, Icons.Filled.Home, "홈"),
+        TabItem(Icons.Outlined.History, Icons.Filled.History, "기록"),
+        TabItem(Icons.Outlined.Analytics, Icons.Filled.Assessment, "분석"),
+        TabItem(Icons.Outlined.Settings, Icons.Filled.Settings, "설정")
+    )
+
+    Surface(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(80.dp),
+        color = MaterialTheme.colorScheme.surface.copy(alpha = 0.95f),
+        tonalElevation = 3.dp
+    ) {
+        // iOS 스타일 블러 효과 배경
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .background(
+                    brush = Brush.verticalGradient(
+                        colors = listOf(
+                            Color.Transparent,
+                            MaterialTheme.colorScheme.surface.copy(alpha = 0.1f)
+                        )
+                    )
+                )
+        ) {
+            Row(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(horizontal = 8.dp, vertical = 8.dp),
+                horizontalArrangement = Arrangement.SpaceEvenly,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                tabs.forEachIndexed { index, tab ->
+                    BottomBarTab(
+                        tab = tab,
+                        isSelected = selectedTab == index,
+                        onClick = { onTabSelected(index) },
+                        modifier = Modifier.weight(1f)
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+private fun BottomBarTab(
+    tab: TabItem,
+    isSelected: Boolean,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val animatedScale by animateFloatAsState(
+        targetValue = if (isSelected) 1.1f else 1f,
+        animationSpec = spring(
+            dampingRatio = Spring.DampingRatioMediumBouncy,
+            stiffness = Spring.StiffnessLow
+        )
+    )
+
+    val animatedColor by animateColorAsState(
+        targetValue = if (isSelected) Primary700 else Gray700,
+        animationSpec = tween(300)
+    )
+
+    Surface(
+        onClick = onClick,
+        modifier = modifier
+            .scale(animatedScale)
+            .padding(4.dp),
+        color = Color.Transparent,
+        shape = RoundedCornerShape(16.dp)
+    ) {
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .background(
+                    if (isSelected) {
+                        Brush.radialGradient(
+                            colors = listOf(
+                                Primary700.copy(alpha = 0.1f),
+                                Color.Transparent
+                            )
+                        )
+                    } else {
+                        Brush.radialGradient(
+                            colors = listOf(Color.Transparent, Color.Transparent)
+                        )
+                    }
+                ),
+            contentAlignment = Alignment.Center
+        ) {
+            Column(
+                horizontalAlignment = Alignment.CenterHorizontally,
+                verticalArrangement = Arrangement.Center,
+                modifier = Modifier.padding(vertical = 4.dp)
+            ) {
+                Icon(
+                    imageVector = if (isSelected) tab.selectedIcon else tab.icon,
+                    contentDescription = tab.label,
+                    tint = animatedColor,
+                    modifier = Modifier.size(24.dp)
+                )
+
+                Spacer(modifier = Modifier.height(4.dp))
+
+                Text(
+                    text = tab.label,
+                    style = MaterialTheme.typography.labelSmall,
+                    color = animatedColor,
+                    fontWeight = if (isSelected) FontWeight.SemiBold else FontWeight.Normal
+                )
+            }
+        }
+    }
+}
+
+private data class TabItem(
+    val icon: ImageVector,
+    val selectedIcon: ImageVector,
+    val label: String
+)
